LPS-200088 Remove the interface and impl by merging the logic of the impl to the Util class and then move the util class PortletContextFactoryUtil to portal-impl as it is only used in portal-impl
# breaking_change_report

## What portal-kernel/src/com/liferay/portal/kernel/portlet/PortletContextFactoryUtil.java portal-kernel/src/com/liferay/portal/kernel/portlet/PortletContextFactory.java portal-impl/src/com/liferay/portlet/internal/PortletContextFactoryImpl.java

PortletContextFactory and PortletContextFactoryImpl are removed and PortletContextFactoryUtil is moving from portal-kernel into portal-impl.

## Why

We are merging portal-kernel into portal-impl as a long term goal to simplify portal structure.

## Alternatives

Add portal-impl in build dependency and use PortletContextFactoryUtil.
